Ensure shebang lines are correctly quoted on Windows. Fixes issue #398.
The executable path was passed as a single string which, if it contained
spaces, caused it to be parsed as a command with a series of arguments,
rather than as a single command containing spaces. As a consequence, the
path in the generated shebang line wasn't quoted which caused the
generated exe wrapper to trip over itself. By passing the executable path
as a single-argument list, paths with spaces aren't split and are quoted
properly.